feat: add UserService interface
New layer of abstraction, encapsulation, flexibility and
separation of concerns
