+++ src/main/java/com/exercises/library/controller/ImageController.java
@@ -1,6 +1,6 @@
 package com.exercises.library.controller;
 
-import com.exercises.library.service.UserService;
+import com.exercises.library.service.UserServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -14,16 +14,16 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @Controller
 @RequestMapping("/image")
 public class ImageController {
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
     @Autowired
-    public ImageController(UserService userService) {
-        this.userService = userService;
+    public ImageController(UserServiceImpl userServiceImpl) {
+        this.userServiceImpl = userServiceImpl;
     }
 
     @GetMapping("/perfil/{id}")
     public ResponseEntity<byte []> userImage(@PathVariable String id) {
-        byte[] image = userService.getUserById(id).getImage().getContent();
+        byte[] image = userServiceImpl.getUserById(id).getImage().getContent();
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.IMAGE_JPEG);
         return new ResponseEntity<>(image, headers, HttpStatus.OK);
